Remove `Completable` from all `CompletableFuture` subclasses for brevity, as suggested by @minwoox
