+++ core/src/main/java/com/linecorp/armeria/internal/RequestContextAwareFuture.java
@@ -25,77 +25,74 @@ import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
 
 import com.linecorp.armeria.common.RequestContext;
 
-final class Java9RequestContextAwareCompletableFuture<T>
-        extends AbstractRequestContextAwareCompletableFuture<T> {
+final class RequestContextAwareFuture<T> extends AbstractRequestContextAwareFuture<T> {
 
-    Java9RequestContextAwareCompletableFuture(RequestContext requestContext) {
+    RequestContextAwareFuture(RequestContext requestContext) {
         super(requestContext);
     }
 
     @Override
     public <U> CompletableFuture<U> thenApply(Function<? super T, ? extends U> fn) {
-        return super.thenApply(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenApply(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> thenApplyAsync(Function<? super T, ? extends U> fn) {
-        return super.thenApplyAsync(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenApplyAsync(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> thenApplyAsync(Function<? super T, ? extends U> fn, Executor executor) {
-        requireNonNull(executor, "executor");
-        return super.thenApplyAsync(makeContextAwareLoggingException(fn), executor);
+        return ctx().makeContextAware(super.thenApplyAsync(makeContextAwareLoggingException(fn), executor));
     }
 
     @Override
     public CompletableFuture<Void> thenAccept(Consumer<? super T> action) {
-        return super.thenAccept(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.thenAccept(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> thenAcceptAsync(Consumer<? super T> action) {
-        return super.thenAcceptAsync(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.thenAcceptAsync(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> thenAcceptAsync(Consumer<? super T> action, Executor executor) {
-        requireNonNull(executor, "executor");
-        return super.thenAcceptAsync(makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.thenAcceptAsync(makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public CompletableFuture<Void> thenRun(Runnable action) {
-        return super.thenRun(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.thenRun(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> thenRunAsync(Runnable action) {
-        return super.thenRunAsync(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.thenRunAsync(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> thenRunAsync(Runnable action, Executor executor) {
         requireNonNull(executor, "executor");
-        return super.thenRunAsync(makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(super.thenRunAsync(makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public <U, V> CompletableFuture<V> thenCombine(CompletionStage<? extends U> other,
                                                    BiFunction<? super T, ? super U, ? extends V> fn) {
         requireNonNull(other, "other");
-        return super.thenCombine(other, makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenCombine(other, makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U, V> CompletableFuture<V> thenCombineAsync(CompletionStage<? extends U> other,
                                                         BiFunction<? super T, ? super U, ? extends V> fn) {
         requireNonNull(other, "other");
-        return super.thenCombineAsync(other, makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenCombineAsync(other, makeContextAwareLoggingException(fn)));
     }
 
     @Override
@@ -104,21 +101,23 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                         Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.thenCombineAsync(other, makeContextAwareLoggingException(fn), executor);
+        return ctx().makeContextAware(
+                super.thenCombineAsync(other, makeContextAwareLoggingException(fn), executor));
     }
 
     @Override
     public <U> CompletableFuture<Void> thenAcceptBoth(CompletionStage<? extends U> other,
                                                       BiConsumer<? super T, ? super U> action) {
         requireNonNull(other, "other");
-        return super.thenAcceptBoth(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.thenAcceptBoth(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
     public <U> CompletableFuture<Void> thenAcceptBothAsync(CompletionStage<? extends U> other,
                                                            BiConsumer<? super T, ? super U> action) {
         requireNonNull(other, "other");
-        return super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(
+                super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
@@ -127,19 +126,23 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                            Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action), executor));
     }
 
     @Override
-    public CompletableFuture<Void> runAfterBoth(CompletionStage<?> other, Runnable action) {
+    public CompletableFuture<Void> runAfterBoth(CompletionStage<?> other,
+                                                Runnable action) {
         requireNonNull(other, "other");
-        return super.runAfterBoth(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.runAfterBoth(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
-    public CompletableFuture<Void> runAfterBothAsync(CompletionStage<?> other, Runnable action) {
+    public CompletableFuture<Void> runAfterBothAsync(CompletionStage<?> other,
+                                                     Runnable action) {
         requireNonNull(other, "other");
-        return super.runAfterBothAsync(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(
+                super.runAfterBothAsync(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
@@ -148,21 +151,22 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                      Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.runAfterBothAsync(other, makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.runAfterBothAsync(other, makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public <U> CompletableFuture<U> applyToEither(CompletionStage<? extends T> other,
                                                   Function<? super T, U> fn) {
         requireNonNull(other, "other");
-        return super.applyToEither(other, makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.applyToEither(other, makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> applyToEitherAsync(CompletionStage<? extends T> other,
                                                        Function<? super T, U> fn) {
         requireNonNull(other, "other");
-        return super.applyToEitherAsync(other, makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.applyToEitherAsync(other, makeContextAwareLoggingException(fn)));
     }
 
     @Override
@@ -171,21 +175,23 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                        Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.applyToEitherAsync(other, makeContextAwareLoggingException(fn), executor);
+        return ctx().makeContextAware(
+                super.applyToEitherAsync(other, makeContextAwareLoggingException(fn), executor));
     }
 
     @Override
     public CompletableFuture<Void> acceptEither(CompletionStage<? extends T> other,
                                                 Consumer<? super T> action) {
         requireNonNull(other, "other");
-        return super.acceptEither(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.acceptEither(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> acceptEitherAsync(CompletionStage<? extends T> other,
                                                      Consumer<? super T> action) {
         requireNonNull(other, "other");
-        return super.acceptEitherAsync(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(
+                super.acceptEitherAsync(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
@@ -194,19 +200,21 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                      Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.acceptEitherAsync(other, makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.acceptEitherAsync(other, makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public CompletableFuture<Void> runAfterEither(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return super.runAfterEither(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.runAfterEither(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<Void> runAfterEitherAsync(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return super.runAfterEitherAsync(other, makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(
+                super.runAfterEitherAsync(other, makeContextAwareLoggingException(action)));
     }
 
     @Override
@@ -215,83 +223,64 @@ final class Java9RequestContextAwareCompletableFuture<T>
                                                        Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return super.runAfterEitherAsync(other, makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.runAfterEitherAsync(other, makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public <U> CompletableFuture<U> thenCompose(Function<? super T, ? extends CompletionStage<U>> fn) {
-        return super.thenCompose(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenCompose(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> thenComposeAsync(Function<? super T, ? extends CompletionStage<U>> fn) {
-        return super.thenComposeAsync(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.thenComposeAsync(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> thenComposeAsync(Function<? super T, ? extends CompletionStage<U>> fn,
                                                      Executor executor) {
         requireNonNull(executor, "executor");
-        return super.thenComposeAsync(makeContextAwareLoggingException(fn), executor);
+        return ctx().makeContextAware(super.thenComposeAsync(makeContextAwareLoggingException(fn), executor));
     }
 
     @Override
     public CompletableFuture<T> whenComplete(BiConsumer<? super T, ? super Throwable> action) {
-        return super.whenComplete(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.whenComplete(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<T> whenCompleteAsync(BiConsumer<? super T, ? super Throwable> action) {
-        return super.whenCompleteAsync(makeContextAwareLoggingException(action));
+        return ctx().makeContextAware(super.whenCompleteAsync(makeContextAwareLoggingException(action)));
     }
 
     @Override
     public CompletableFuture<T> whenCompleteAsync(BiConsumer<? super T, ? super Throwable> action,
                                                   Executor executor) {
         requireNonNull(executor, "executor");
-        return super.whenCompleteAsync(makeContextAwareLoggingException(action), executor);
+        return ctx().makeContextAware(
+                super.whenCompleteAsync(makeContextAwareLoggingException(action), executor));
     }
 
     @Override
     public <U> CompletableFuture<U> handle(BiFunction<? super T, Throwable, ? extends U> fn) {
-        return super.handle(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.handle(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> handleAsync(BiFunction<? super T, Throwable, ? extends U> fn) {
-        return super.handleAsync(makeContextAwareLoggingException(fn));
+        return ctx().makeContextAware(super.handleAsync(makeContextAwareLoggingException(fn)));
     }
 
     @Override
     public <U> CompletableFuture<U> handleAsync(BiFunction<? super T, Throwable, ? extends U> fn,
                                                 Executor executor) {
         requireNonNull(executor, "executor");
-        return super.handleAsync(makeContextAwareLoggingException(fn), executor);
+        return ctx().makeContextAware(super.handleAsync(makeContextAwareLoggingException(fn), executor));
     }
 
     @Override
     public CompletableFuture<T> exceptionally(Function<Throwable, ? extends T> fn) {
-        return super.exceptionally(makeContextAwareLoggingException(fn));
-    }
-
-    @Override
-    public <U> CompletableFuture<U> newIncompleteFuture() {
-        return new Java9RequestContextAwareCompletableFuture<>(ctx());
-    }
-
-    @Override
-    public CompletionStage<T> minimalCompletionStage() {
-        return new Java9RequestContextAwareMinimalStage<>(this);
-    }
-
-    @Override
-    public CompletableFuture<T> completeAsync(Supplier<? extends T> supplier) {
-        return super.completeAsync(makeContextAwareLoggingException(supplier));
-    }
-
-    @Override
-    public CompletableFuture<T> completeAsync(Supplier<? extends T> supplier, Executor executor) {
-        requireNonNull(executor, "executor");
-        return super.completeAsync(makeContextAwareLoggingException(supplier), executor);
+        return ctx().makeContextAware(super.exceptionally(makeContextAwareLoggingException(fn)));
     }
 }
