+++ core/src/main/java9/com/linecorp/armeria/internal/Java9RequestContextAwareFuture.java
@@ -25,74 +25,77 @@ import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
 import java.util.function.Function;
+import java.util.function.Supplier;
 
 import com.linecorp.armeria.common.RequestContext;
 
-final class RequestContextAwareCompletableFuture<T> extends AbstractRequestContextAwareCompletableFuture<T> {
+final class Java9RequestContextAwareFuture<T>
+        extends AbstractRequestContextAwareFuture<T> {
 
-    RequestContextAwareCompletableFuture(RequestContext requestContext) {
+    Java9RequestContextAwareFuture(RequestContext requestContext) {
         super(requestContext);
     }
 
     @Override
     public <U> CompletableFuture<U> thenApply(Function<? super T, ? extends U> fn) {
-        return ctx().makeContextAware(super.thenApply(makeContextAwareLoggingException(fn)));
+        return super.thenApply(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> thenApplyAsync(Function<? super T, ? extends U> fn) {
-        return ctx().makeContextAware(super.thenApplyAsync(makeContextAwareLoggingException(fn)));
+        return super.thenApplyAsync(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> thenApplyAsync(Function<? super T, ? extends U> fn, Executor executor) {
-        return ctx().makeContextAware(super.thenApplyAsync(makeContextAwareLoggingException(fn), executor));
+        requireNonNull(executor, "executor");
+        return super.thenApplyAsync(makeContextAwareLoggingException(fn), executor);
     }
 
     @Override
     public CompletableFuture<Void> thenAccept(Consumer<? super T> action) {
-        return ctx().makeContextAware(super.thenAccept(makeContextAwareLoggingException(action)));
+        return super.thenAccept(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> thenAcceptAsync(Consumer<? super T> action) {
-        return ctx().makeContextAware(super.thenAcceptAsync(makeContextAwareLoggingException(action)));
+        return super.thenAcceptAsync(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> thenAcceptAsync(Consumer<? super T> action, Executor executor) {
-        return ctx().makeContextAware(
-                super.thenAcceptAsync(makeContextAwareLoggingException(action), executor));
+        requireNonNull(executor, "executor");
+        return super.thenAcceptAsync(makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public CompletableFuture<Void> thenRun(Runnable action) {
-        return ctx().makeContextAware(super.thenRun(makeContextAwareLoggingException(action)));
+        return super.thenRun(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> thenRunAsync(Runnable action) {
-        return ctx().makeContextAware(super.thenRunAsync(makeContextAwareLoggingException(action)));
+        return super.thenRunAsync(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> thenRunAsync(Runnable action, Executor executor) {
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(super.thenRunAsync(makeContextAwareLoggingException(action), executor));
+        return super.thenRunAsync(makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public <U, V> CompletableFuture<V> thenCombine(CompletionStage<? extends U> other,
                                                    BiFunction<? super T, ? super U, ? extends V> fn) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.thenCombine(other, makeContextAwareLoggingException(fn)));
+        return super.thenCombine(other, makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U, V> CompletableFuture<V> thenCombineAsync(CompletionStage<? extends U> other,
                                                         BiFunction<? super T, ? super U, ? extends V> fn) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.thenCombineAsync(other, makeContextAwareLoggingException(fn)));
+        return super.thenCombineAsync(other, makeContextAwareLoggingException(fn));
     }
 
     @Override
@@ -101,23 +104,21 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                         Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.thenCombineAsync(other, makeContextAwareLoggingException(fn), executor));
+        return super.thenCombineAsync(other, makeContextAwareLoggingException(fn), executor);
     }
 
     @Override
     public <U> CompletableFuture<Void> thenAcceptBoth(CompletionStage<? extends U> other,
                                                       BiConsumer<? super T, ? super U> action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.thenAcceptBoth(other, makeContextAwareLoggingException(action)));
+        return super.thenAcceptBoth(other, makeContextAwareLoggingException(action));
     }
 
     @Override
     public <U> CompletableFuture<Void> thenAcceptBothAsync(CompletionStage<? extends U> other,
                                                            BiConsumer<? super T, ? super U> action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(
-                super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action)));
+        return super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action));
     }
 
     @Override
@@ -126,23 +127,19 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                            Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action), executor));
+        return super.thenAcceptBothAsync(other, makeContextAwareLoggingException(action), executor);
     }
 
     @Override
-    public CompletableFuture<Void> runAfterBoth(CompletionStage<?> other,
-                                                Runnable action) {
+    public CompletableFuture<Void> runAfterBoth(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.runAfterBoth(other, makeContextAwareLoggingException(action)));
+        return super.runAfterBoth(other, makeContextAwareLoggingException(action));
     }
 
     @Override
-    public CompletableFuture<Void> runAfterBothAsync(CompletionStage<?> other,
-                                                     Runnable action) {
+    public CompletableFuture<Void> runAfterBothAsync(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(
-                super.runAfterBothAsync(other, makeContextAwareLoggingException(action)));
+        return super.runAfterBothAsync(other, makeContextAwareLoggingException(action));
     }
 
     @Override
@@ -151,22 +148,21 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                      Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.runAfterBothAsync(other, makeContextAwareLoggingException(action), executor));
+        return super.runAfterBothAsync(other, makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public <U> CompletableFuture<U> applyToEither(CompletionStage<? extends T> other,
                                                   Function<? super T, U> fn) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.applyToEither(other, makeContextAwareLoggingException(fn)));
+        return super.applyToEither(other, makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> applyToEitherAsync(CompletionStage<? extends T> other,
                                                        Function<? super T, U> fn) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.applyToEitherAsync(other, makeContextAwareLoggingException(fn)));
+        return super.applyToEitherAsync(other, makeContextAwareLoggingException(fn));
     }
 
     @Override
@@ -175,23 +171,21 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                        Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.applyToEitherAsync(other, makeContextAwareLoggingException(fn), executor));
+        return super.applyToEitherAsync(other, makeContextAwareLoggingException(fn), executor);
     }
 
     @Override
     public CompletableFuture<Void> acceptEither(CompletionStage<? extends T> other,
                                                 Consumer<? super T> action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.acceptEither(other, makeContextAwareLoggingException(action)));
+        return super.acceptEither(other, makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> acceptEitherAsync(CompletionStage<? extends T> other,
                                                      Consumer<? super T> action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(
-                super.acceptEitherAsync(other, makeContextAwareLoggingException(action)));
+        return super.acceptEitherAsync(other, makeContextAwareLoggingException(action));
     }
 
     @Override
@@ -200,21 +194,19 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                      Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.acceptEitherAsync(other, makeContextAwareLoggingException(action), executor));
+        return super.acceptEitherAsync(other, makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public CompletableFuture<Void> runAfterEither(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(super.runAfterEither(other, makeContextAwareLoggingException(action)));
+        return super.runAfterEither(other, makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<Void> runAfterEitherAsync(CompletionStage<?> other, Runnable action) {
         requireNonNull(other, "other");
-        return ctx().makeContextAware(
-                super.runAfterEitherAsync(other, makeContextAwareLoggingException(action)));
+        return super.runAfterEitherAsync(other, makeContextAwareLoggingException(action));
     }
 
     @Override
@@ -223,64 +215,83 @@ final class RequestContextAwareCompletableFuture<T> extends AbstractRequestConte
                                                        Executor executor) {
         requireNonNull(other, "other");
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.runAfterEitherAsync(other, makeContextAwareLoggingException(action), executor));
+        return super.runAfterEitherAsync(other, makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public <U> CompletableFuture<U> thenCompose(Function<? super T, ? extends CompletionStage<U>> fn) {
-        return ctx().makeContextAware(super.thenCompose(makeContextAwareLoggingException(fn)));
+        return super.thenCompose(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> thenComposeAsync(Function<? super T, ? extends CompletionStage<U>> fn) {
-        return ctx().makeContextAware(super.thenComposeAsync(makeContextAwareLoggingException(fn)));
+        return super.thenComposeAsync(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> thenComposeAsync(Function<? super T, ? extends CompletionStage<U>> fn,
                                                      Executor executor) {
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(super.thenComposeAsync(makeContextAwareLoggingException(fn), executor));
+        return super.thenComposeAsync(makeContextAwareLoggingException(fn), executor);
     }
 
     @Override
     public CompletableFuture<T> whenComplete(BiConsumer<? super T, ? super Throwable> action) {
-        return ctx().makeContextAware(super.whenComplete(makeContextAwareLoggingException(action)));
+        return super.whenComplete(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<T> whenCompleteAsync(BiConsumer<? super T, ? super Throwable> action) {
-        return ctx().makeContextAware(super.whenCompleteAsync(makeContextAwareLoggingException(action)));
+        return super.whenCompleteAsync(makeContextAwareLoggingException(action));
     }
 
     @Override
     public CompletableFuture<T> whenCompleteAsync(BiConsumer<? super T, ? super Throwable> action,
                                                   Executor executor) {
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(
-                super.whenCompleteAsync(makeContextAwareLoggingException(action), executor));
+        return super.whenCompleteAsync(makeContextAwareLoggingException(action), executor);
     }
 
     @Override
     public <U> CompletableFuture<U> handle(BiFunction<? super T, Throwable, ? extends U> fn) {
-        return ctx().makeContextAware(super.handle(makeContextAwareLoggingException(fn)));
+        return super.handle(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> handleAsync(BiFunction<? super T, Throwable, ? extends U> fn) {
-        return ctx().makeContextAware(super.handleAsync(makeContextAwareLoggingException(fn)));
+        return super.handleAsync(makeContextAwareLoggingException(fn));
     }
 
     @Override
     public <U> CompletableFuture<U> handleAsync(BiFunction<? super T, Throwable, ? extends U> fn,
                                                 Executor executor) {
         requireNonNull(executor, "executor");
-        return ctx().makeContextAware(super.handleAsync(makeContextAwareLoggingException(fn), executor));
+        return super.handleAsync(makeContextAwareLoggingException(fn), executor);
     }
 
     @Override
     public CompletableFuture<T> exceptionally(Function<Throwable, ? extends T> fn) {
-        return ctx().makeContextAware(super.exceptionally(makeContextAwareLoggingException(fn)));
+        return super.exceptionally(makeContextAwareLoggingException(fn));
+    }
+
+    @Override
+    public <U> CompletableFuture<U> newIncompleteFuture() {
+        return new Java9RequestContextAwareFuture<>(ctx());
+    }
+
+    @Override
+    public CompletionStage<T> minimalCompletionStage() {
+        return new Java9RequestContextAwareMinimalStage<>(this);
+    }
+
+    @Override
+    public CompletableFuture<T> completeAsync(Supplier<? extends T> supplier) {
+        return super.completeAsync(makeContextAwareLoggingException(supplier));
+    }
+
+    @Override
+    public CompletableFuture<T> completeAsync(Supplier<? extends T> supplier, Executor executor) {
+        requireNonNull(executor, "executor");
+        return super.completeAsync(makeContextAwareLoggingException(supplier), executor);
     }
 }
