+++ core/src/test/java9/com/linecorp/armeria/internal/Java9RequestContextAwareFutureTest.java
@@ -58,7 +58,7 @@ import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.Appender;
 
-class Java9RequestContextAwareCompletableFutureTest {
+class Java9RequestContextAwareFutureTest {
 
     // TODO(minwoox) Make an extesion which a user can easily check the logs.
     @Mock
