+++ thrift/src/test/java/com/linecorp/armeria/common/thrift/ThriftFutureTest.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2020 LINE Corporation
+ *
+ * LINE Corporation licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package com.linecorp.armeria.common.thrift;
+
+import static com.linecorp.armeria.common.thrift.ThriftFuture.completedFuture;
+import static com.linecorp.armeria.common.thrift.ThriftFuture.exceptionallyCompletedFuture;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.Test;
+
+class ThriftFutureTest {
+
+    @Test
+    void testSuccessfulCompletedFuture() throws Exception {
+        final ThriftFuture<String> future = completedFuture("success");
+        assertThat(future.get()).isEqualTo("success");
+    }
+
+    @Test
+    void testFailedCompletedFuture() throws Exception {
+        final ThriftFuture<String> future = exceptionallyCompletedFuture(new IllegalStateException());
+        assertThatThrownBy(future::get).hasCauseInstanceOf(IllegalStateException.class);
+    }
+}
